+++ src/main/Game.java
@@ -2,8 +2,10 @@ package main;
 
 import java.awt.Graphics;
 import java.awt.Toolkit;
+import java.util.logging.Level;
 
 import entities.Player;
+import levels.LevelManager;
 
 public class Game implements Runnable {
 
@@ -14,6 +16,15 @@ public class Game implements Runnable {
 	private final int UPS_SET = 200;
 
 	private Player player;
+	private LevelManager levelManager;
+
+	public final static int TILES_DEFAULT_SIZE = 24;
+	public final static float SCALE = 2f; 
+	public final static int TILES_IN_WIDTH = 26;
+	public final static int TILES_IN_HEIGHT = 14;
+	public final static int TILES_SIZE = (int) (TILES_DEFAULT_SIZE * SCALE);
+	public final static int GAME_WIDTH = TILES_SIZE * TILES_IN_WIDTH;
+	public final static int GAME_HEIGHT = TILES_SIZE * TILES_IN_HEIGHT;
 
 	public Game() {
 
@@ -26,6 +37,7 @@ public class Game implements Runnable {
 	}
 
 	private void initClasses() {
+		levelManager = new LevelManager(this);
 		player = new Player(200, 200);
 	}
 
@@ -36,9 +48,11 @@ public class Game implements Runnable {
 
 	private void update() {
 		player.update();
+		levelManager.update();
 	}
 
 	public void render(Graphics g) {
+		levelManager.draw(g);
 		player.render(g);
 	}